new library
rewriting the program code using the vkbottle library
