+++ handlers/__init__.py
@@ -1,0 +1,5 @@
+from . import comand
+
+labelers = [comand.labeler]
+
+__all__ = ["labelers"]
