+++ handlers/reader.py
@@ -1,0 +1,3 @@
+async def read_file(file_path):
+    with open(file_path, "r", encoding="utf-8") as f:
+        return f.read()
